ilook_v1/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
class-registry==2.1.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.5
django-filter==25.1
django-environ==0.12.0
djangorestframework==3.16.1
docopt==0.6.2
et_xmlfile==2.0.0
filters==1.3.2
filters-django==1.0.5
idna==3.10
kombu==5.5.4
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pipreqs==0.4.13
prompt_toolkit==3.0.51
psycopg==3.2.9
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
regex==2025.7.34
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
yarg==0.1.10